+++ test/wrap_widget.dart
@@ -20,6 +20,12 @@ class WrapWidgetState extends State<WrapWidget> {
   FlutterReactionType? flutterReactionType;
 
   @override
+  void dispose() {
+    context.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     return FlutterReactionButton(
       value: flutterReactionType,